save: independently limit memory usage of various stages.
We want high parallelism for the final stage of writing to GCS, since
we're forced to do so in order to have high bandwidth given the high
latency. But we don't need that high parallelism when reading for disk
and encrypting, and indeed having it causes overly high memory usage.